Dumping some Slider class code, WIP!

+++ include/ulcdgui/rendering/color.h
@@ -13,7 +13,7 @@ protected:
 public:
     Color();
     Color(int r, int g, int b);
-    Color(float r, float g, float b);
+    Color(double r, double g, double b);
     Color(uint16_t c);
 
     float getR() { return this->r; }
